feat: Add create_action function to save user actions
The code changes add the `create_action` function to the `images/views.py` file. This function is used to save user actions, such as bookmarking an image or liking an image. It takes the user, action type, and the image as parameters and creates a new `Action` object to store the action in the database. This functionality enhances the tracking and display of user actions within the application.