feat: Optimize image_detail view with prefetch_related and select_related
The code changes in the `images/views.py` file optimize the `image_detail` view by using `prefetch_related` and `select_related` to fetch related objects in a more efficient manner. This improves the performance of the view by reducing the number of database queries and enhances the overall user experience.