feat: Add constant for action delay in actions module
The code changes in the `actions/constants.py` file add a new constant `ACTION_DELAY` with a value of 60 seconds. This constant is used in the `actions/utils.py` file to set the time window for filtering similar actions. By introducing this constant, the code becomes more maintainable and allows for easy adjustment of the action delay in the future.