feat: Add total_likes field to Image model and create index
The code changes add a new field, `total_likes`, to the `Image` model in the `images/models.py` file. This field is a `PositiveIntegerField` with a default value of 0. Additionally, an index is created on the `total_likes` field to optimize queries that involve sorting or filtering based on the total number of likes. This enhancement improves the functionality and performance of the application.