feat: Add language selection dropdown to header
This commit modifies the base.html template to add a language selection dropdown to the header of the shop. The dropdown is populated with available languages and allows users to switch between different language versions of the site. This feature enhances the localization support of the application and improves the user experience.