feat: Add django-parler for multilingual support
This commit adds the 'parler' package to the list of installed apps in the Django settings. It also includes the configuration for multilingual support using the 'parler' package. The 'PARLER_LANGUAGES' setting is updated to include English and Spanish as available languages. This feature enhances the localization support of the application and allows for easy translation of the user interface into different languages.