feat: Add formset.html template for editing course modules
This commit adds the formset.html template to the courses app, specifically in the `courses/manage/module` directory. The template extends the base.html template and provides a form for editing course modules. It includes a heading displaying the title of the course being edited, a section for displaying the course modules, and a form for saving the changes. This code change enhances the functionality of the application by allowing users to easily edit and save course modules.