feat: Add form template for creating/editing courses
This commit adds the form.html template to the courses app, specifically in the `courses/manage/course` directory. The template extends the base.html template and provides a form for creating or editing a course. It includes fields for the course title and a submit button to save the changes. This code change enhances the functionality of the application by allowing users to easily create and edit courses.