feat: Add inline formset for managing course modules
This code change adds an inline formset for managing course modules in the `Course` model. The `ModuleFormSet` is created using the `inlineformset_factory` function from Django's form module. It allows users to add, edit, and delete modules for a course. This enhancement improves the functionality of the application by providing a convenient way to manage course modules.