feat: Add template tag for getting model name
This code change adds a template tag called `model_name` to the `course` templatetags module. The `model_name` template tag takes an object as input and returns the model name of the object. It first tries to access the `_meta.model_name` attribute of the object, and if it doesn't exist, it returns `None`. This enhancement improves the functionality of the application by providing a convenient way to retrieve the model name of an object in templates.