feat: Add manage contents link to course list page
This code change adds a new link to the course list page that allows users to manage the contents of a course. The link is only displayed if the course has at least one module. This enhancement improves the user experience by providing a convenient way to access and manage course contents directly from the course list page.