feat: Add URLs for course list and course detail views
This code change adds new URLs for the course list and course detail views in the courses app. The course list view, CourseListView, is responsible for displaying and filtering courses based on subjects. It retrieves the subjects and courses from the database and allows users to filter courses by subject. The course detail view, CourseDetailView, is responsible for displaying the details of a specific course. It retrieves the course information from the database and renders it using the 'courses/course/detail.html' template. These enhancements improve the functionality of the application by providing dedicated views for listing, filtering, and displaying course details.