feat: Add students field to Course model
This code change adds a new field, 'students', to the Course model in the courses app. The 'students' field is a ManyToManyField that allows for a many-to-many relationship between the Course model and the User model. It is used to track the students who have joined a particular course. This enhancement improves the functionality of the application by providing a way to associate students with courses.