feat: Add course list template for students
This code change adds a new template file, `list.html`, to the `students/course` directory. The template is used to display a list of courses for the authenticated user. If the user is not enrolled in any courses, a message and a link to browse available courses are shown. This enhancement improves the user experience by providing a dedicated page for students to view their enrolled courses.