chore: Update Django and other dependencies
This code change updates the Django version to 4.1.13 and other dependencies in the requirements.txt file. By updating these dependencies, we ensure that the application is using the latest stable versions and taking advantage of bug fixes and new features. This chore improves the stability and performance of the application.